switch to PSR-4 based naming
