+++ bin/update-external-strings
@@ -2,8 +2,8 @@
 
 set -e
 
-wp-i18n-export koko-analytics-pro ../koko-analytics-pro/includes > src/external-strings.php
-./vendor/bin/phpcbf src/external-strings.php || true
+wp-i18n-export koko-analytics-pro ../koko-analytics-pro/includes > src/Resources/external-strings.php
+./vendor/bin/phpcbf src/Resources/external-strings.php || true
 
-git add src/external-strings.php || true
+git add src/Resources/external-strings.php || true
 git commit -m "Update external strings from Koko Analytics Pro"
