+++ src/Dashboard.php
@@ -12,7 +12,7 @@ class Dashboard
 {
     public function show_standalone_dashboard_page(): void
     {
-        require __DIR__ . '/views/standalone.php';
+        require KOKO_ANALYTICS_PLUGIN_DIR . '/src/Resources/views/standalone.php';
         exit;
     }
 
@@ -76,7 +76,7 @@ class Dashboard
 
         $chart_data =  $stats->get_stats($dateStart->format("Y-m-d"), $dateEnd->format('Y-m-d'), $groupChartBy, $page);
 
-        require __DIR__ . '/views/dashboard-page.php';
+        require KOKO_ANALYTICS_PLUGIN_DIR . '/src/Resources/views/dashboard-page.php';
     }
 
     public function get_next_period(\DateTimeImmutable $dateStart, \DateTimeImmutable $dateEnd, int $dir = 1): array
@@ -205,6 +205,8 @@ class Dashboard
             case 'all_time':
                 return (new Stats())->get_total_date_range();
         }
+
+        throw new \Exception("invalid preset key: {$key}");
     }
 
     public function get_first_day_of_current_week(\DateTimeImmutable $now, int $week_starts_on = 0): \DateTimeImmutable
