+++ src/Resources/functions/global.php
