+++ src/Resources/views/standalone.php
