--- src/external-strings.php
@@ -1,942 +1,0 @@
-<?php
-
-// i18n strings for koko-analytics-pro
-
-// ../koko-analytics-pro/includes/views/geolocation-settings.php#7
-__('Enable geo-location?', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/geolocation-settings.php#12
-__('Select "yes" if you want Koko Analytics to geo-locate visitors by their IP address.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-country-component.php#12
-__('Country', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-country-component.php#13
-__('Count', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-country-component.php#30
-__('There is nothing here. Yet!', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-country-component.php#40
-__('Previous', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-country-component.php#43
-__('Next', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-event-component.php#14
-__('Event:', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-event-component.php#15
-__('Unique', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-event-component.php#16
-__('Total', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-event-component.php#42
-__('There is nothing here. Yet!', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-event-component.php#52
-__('Previous', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/dashboard-event-component.php#55
-__('Next', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/traffic-spike-notification-settings.php#6
-__('Traffic Spike Notification', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/traffic-spike-notification-settings.php#9
-__('Send email notification of traffic spikes?', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/traffic-spike-notification-settings.php#16
-__('Realtime pageview treshold', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/traffic-spike-notification-settings.php#18
-__('The notification will be sent if there were more pageviews than the specified treshold within a single hour.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/traffic-spike-notification-settings.php#22
-__('Send to these email addresses', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/traffic-spike-notification-settings.php#24
-__('Enter a comma separated list of email addresses to which the notification should be sent.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/email-reports-setting.php#6
-__('Email Reports', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/email-reports-setting.php#9
-__('Send periodic email reports?', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/email-reports-setting.php#11
-__('Daily', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/email-reports-setting.php#11
-__('Weekly', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/email-reports-setting.php#11
-__('Monthly', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/email-reports-setting.php#17
-__('Select the timeframes for which you want to receive a summary of your most important statistics.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/email-reports-setting.php#21
-__('Send to these email addresses', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/email-reports-setting.php#23
-__('Enter a comma separated list of email addresses to which the reports should be sent.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#10
-__('Events', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#17
-__('Track outbound link clicks?', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#21
-__('Select "yes" if you want Koko Analytics to count all clicks on links to external websites.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#26
-__('Track form submissions?', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#30
-__('Select "yes" if you want Koko Analytics to count all form submissions.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#39
-__('Event types', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#40
-__('You currently have the following event types registered:', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#43
-__('Event name', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#54
-__('You have not yet registered any events.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#61
-__('Add new event type', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#62
-__('Use the form below to register new event types.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#65
-__('Event name', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/events-section.php#67
-__('Add event', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#27
-__('Event', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#30
-__('This is a built-in event type that counts all clicks on outbound links on your site.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#33
-__('This is a built-in event type that counts all form submissions on your site.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#38
-__('This event has been counted %1$s times for a total value of %2$s.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#40
-__('This event has been counted %1$s times.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#45
-__('Tracking events', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#46
-__('To track an event of this type, run the following JavaScript code:', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#49
-__('You can optionally attach a numerical value to each recorded event:', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#52
-__('You can find more information on tracking events in our knowledge base.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#56
-__('Export this event', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#57
-__('Use the form below to export event counts in a specific time period to CSV.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#61
-__('Start date', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#65
-__('End date', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#68
-__('Export to CSV', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#74
-__('Delete this event', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#75
-__('Use the button below to delete this event type from your dashboard overview.', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#76
-__('Are you sure you want to delete this event?', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#78
-__('Delete event', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/views/event-details.php#84
-__('Back to settings page', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-admin-bar.php#57
-__('Pageviews', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#12
-__('Bangladesh', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#13
-__('Belgium', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#14
-__('Burkina Faso', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#15
-__('Bulgaria', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#16
-__('Bosnia and Herzegovina', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#17
-__('Barbados', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#18
-__('Wallis and Futuna', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#19
-__('Saint Barthelemy', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#20
-__('Bermuda', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#21
-__('Brunei', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#22
-__('Bolivia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#23
-__('Bahrain', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#24
-__('Burundi', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#25
-__('Benin', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#26
-__('Bhutan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#27
-__('Jamaica', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#28
-__('Bouvet Island', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#29
-__('Botswana', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#30
-__('Samoa', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#31
-__('Bonaire, Saint Eustatius and Saba ', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#32
-__('Brazil', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#33
-__('Bahamas', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#34
-__('Jersey', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#35
-__('Belarus', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#36
-__('Belize', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#37
-__('Russia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#38
-__('Rwanda', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#39
-__('Serbia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#40
-__('East Timor', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#41
-__('Reunion', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#42
-__('Turkmenistan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#43
-__('Tajikistan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#44
-__('Romania', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#45
-__('Tokelau', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#46
-__('Guinea-Bissau', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#47
-__('Guam', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#48
-__('Guatemala', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#49
-__('South Georgia and the South Sandwich Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#50
-__('Greece', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#51
-__('Equatorial Guinea', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#52
-__('Guadeloupe', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#53
-__('Japan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#54
-__('Guyana', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#55
-__('Guernsey', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#56
-__('French Guiana', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#57
-__('Georgia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#58
-__('Grenada', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#59
-__('United Kingdom', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#60
-__('Gabon', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#61
-__('El Salvador', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#62
-__('Guinea', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#63
-__('Gambia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#64
-__('Greenland', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#65
-__('Gibraltar', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#66
-__('Ghana', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#67
-__('Oman', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#68
-__('Tunisia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#69
-__('Jordan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#70
-__('Croatia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#71
-__('Haiti', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#72
-__('Hungary', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#73
-__('Hong Kong', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#74
-__('Honduras', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#75
-__('Heard Island and McDonald Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#76
-__('Venezuela', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#77
-__('Puerto Rico', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#78
-__('Palestinian Territory', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#79
-__('Palau', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#80
-__('Portugal', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#81
-__('Svalbard and Jan Mayen', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#82
-__('Paraguay', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#83
-__('Iraq', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#84
-__('Panama', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#85
-__('French Polynesia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#86
-__('Papua New Guinea', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#87
-__('Peru', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#88
-__('Pakistan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#89
-__('Philippines', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#90
-__('Pitcairn', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#91
-__('Poland', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#92
-__('Saint Pierre and Miquelon', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#93
-__('Zambia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#94
-__('Western Sahara', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#95
-__('Estonia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#96
-__('Egypt', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#97
-__('South Africa', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#98
-__('Ecuador', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#99
-__('Italy', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#100
-__('Vietnam', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#101
-__('Solomon Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#102
-__('Ethiopia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#103
-__('Somalia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#104
-__('Zimbabwe', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#105
-__('Saudi Arabia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#106
-__('Spain', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#107
-__('Eritrea', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#108
-__('Montenegro', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#109
-__('Moldova', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#110
-__('Madagascar', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#111
-__('Saint Martin', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#112
-__('Morocco', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#113
-__('Monaco', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#114
-__('Uzbekistan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#115
-__('Myanmar', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#116
-__('Mali', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#117
-__('Macao', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#118
-__('Mongolia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#119
-__('Marshall Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#120
-__('Macedonia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#121
-__('Mauritius', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#122
-__('Malta', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#123
-__('Malawi', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#124
-__('Maldives', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#125
-__('Martinique', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#126
-__('Northern Mariana Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#127
-__('Montserrat', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#128
-__('Mauritania', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#129
-__('Isle of Man', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#130
-__('Uganda', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#131
-__('Tanzania', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#132
-__('Malaysia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#133
-__('Mexico', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#134
-__('Israel', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#135
-__('France', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#136
-__('British Indian Ocean Territory', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#137
-__('Saint Helena', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#138
-__('Finland', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#139
-__('Fiji', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#140
-__('Falkland Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#141
-__('Micronesia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#142
-__('Faroe Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#143
-__('Nicaragua', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#144
-__('Netherlands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#145
-__('Norway', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#146
-__('Namibia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#147
-__('Vanuatu', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#148
-__('New Caledonia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#149
-__('Niger', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#150
-__('Norfolk Island', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#151
-__('Nigeria', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#152
-__('New Zealand', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#153
-__('Nepal', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#154
-__('Nauru', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#155
-__('Niue', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#156
-__('Cook Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#157
-__('Kosovo', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#158
-__('Ivory Coast', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#159
-__('Switzerland', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#160
-__('Colombia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#161
-__('China', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#162
-__('Cameroon', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#163
-__('Chile', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#164
-__('Cocos Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#165
-__('Canada', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#166
-__('Republic of the Congo', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#167
-__('Central African Republic', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#168
-__('Democratic Republic of the Congo', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#169
-__('Czech Republic', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#170
-__('Cyprus', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#171
-__('Christmas Island', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#172
-__('Costa Rica', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#173
-__('Curacao', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#174
-__('Cape Verde', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#175
-__('Cuba', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#176
-__('Swaziland', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#177
-__('Syria', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#178
-__('Sint Maarten', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#179
-__('Kyrgyzstan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#180
-__('Kenya', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#181
-__('South Sudan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#182
-__('Suriname', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#183
-__('Kiribati', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#184
-__('Cambodia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#185
-__('Saint Kitts and Nevis', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#186
-__('Comoros', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#187
-__('Sao Tome and Principe', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#188
-__('Slovakia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#189
-__('South Korea', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#190
-__('Slovenia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#191
-__('North Korea', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#192
-__('Kuwait', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#193
-__('Senegal', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#194
-__('San Marino', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#195
-__('Sierra Leone', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#196
-__('Seychelles', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#197
-__('Kazakhstan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#198
-__('Cayman Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#199
-__('Singapore', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#200
-__('Sweden', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#201
-__('Sudan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#202
-__('Dominican Republic', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#203
-__('Dominica', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#204
-__('Djibouti', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#205
-__('Denmark', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#206
-__('British Virgin Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#207
-__('Germany', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#208
-__('Yemen', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#209
-__('Algeria', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#210
-__('United States', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#211
-__('Uruguay', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#212
-__('Mayotte', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#213
-__('United States Minor Outlying Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#214
-__('Lebanon', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#215
-__('Saint Lucia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#216
-__('Laos', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#217
-__('Tuvalu', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#218
-__('Taiwan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#219
-__('Trinidad and Tobago', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#220
-__('Turkey', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#221
-__('Sri Lanka', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#222
-__('Liechtenstein', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#223
-__('Latvia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#224
-__('Tonga', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#225
-__('Lithuania', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#226
-__('Luxembourg', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#227
-__('Liberia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#228
-__('Lesotho', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#229
-__('Thailand', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#230
-__('French Southern Territories', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#231
-__('Togo', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#232
-__('Chad', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#233
-__('Turks and Caicos Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#234
-__('Libya', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#235
-__('Vatican', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#236
-__('Saint Vincent and the Grenadines', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#237
-__('United Arab Emirates', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#238
-__('Andorra', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#239
-__('Antigua and Barbuda', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#240
-__('Afghanistan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#241
-__('Anguilla', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#242
-__('U.S. Virgin Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#243
-__('Iceland', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#244
-__('Iran', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#245
-__('Armenia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#246
-__('Albania', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#247
-__('Angola', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#248
-__('Antarctica', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#249
-__('American Samoa', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#250
-__('Argentina', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#251
-__('Australia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#252
-__('Austria', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#253
-__('Aruba', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#254
-__('India', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#255
-__('Aland Islands', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#256
-__('Azerbaijan', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#257
-__('Ireland', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#258
-__('Indonesia', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#259
-__('Ukraine', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#260
-__('Qatar', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-country-codes.php#261
-__('Mozambique', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-dashboard.php#17
-__('Download as CSV', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-posts-table.php#90
-__('Number of days to use for pageviews column', 'koko-analytics');
-
-// ../koko-analytics-pro/includes/class-posts-table.php#98
-__('Pageviews', 'koko-analytics');
